requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.2.0
python-dotenv==1.0.1
google-generativeai==0.3.2